requirements.txt
Jinja2
numpy
sympy
Click
xeger
pycparser
ansicolors
flask
python-dotenv
